event-service/requirements.txt
fastapi
uvicorn
kafka-python
SQLAlchemy
psycopg2
alembic
python-dotenv
